Diplom часть 3. Дополнительный сервис Комментарии. • после 2-го ревью. • подправил поиск комментариев для списка событий.

+++ ewm-service/src/main/java/ru/practicum/explore_with_me/dto/event/EventFullDto.java
@@ -79,5 +79,5 @@ public class EventFullDto {
      * Поле не сохраняем в БД. Оно потом вычисляется.
      */
     private Integer views;
-    private Integer comments;
+    private Long comments;
 }
