+++ ewm-service/src/main/java/ru/practicum/explore_with_me/model/Event.java
@@ -121,5 +121,5 @@ public class Event {
     private Integer views;
 
     @Transient
-    private Integer comments;
+    private Long comments;
 }
